+++ cmake/winrc.cmake
@@ -1,4 +1,7 @@
-set(_rc_content "#include <windows.h>
+macro(add_winrc target)
+    cmake_parse_arguments(ARG "" "NAME;DESCRIPTION;COPYRIGHT" "" ${ARGN})
+    
+    set(_rc_content "#include <windows.h>
 
 #ifndef VS_VERSION_INFO
 #define VS_VERSION_INFO 1
@@ -16,11 +19,11 @@ VS_VERSION_INFO VERSIONINFO
        // U.S. English - Windows, Multilingual
        BLOCK \"040904E4\"
        {
-          VALUE \"FileDescription\", STRINGIFY(${RC_DESCRIPTION})
+          VALUE \"FileDescription\", STRINGIFY(${ARG_DESCRIPTION})
           VALUE \"FileVersion\", STRINGIFY(${PROJECT_VERSION})
-          VALUE \"ProductName\", STRINGIFY(${PROJECT_NAME})
+          VALUE \"ProductName\", STRINGIFY(${ARG_NAME})
           VALUE \"ProductVersion\", STRINGIFY(${PROJECT_VERSION})
-          VALUE \"LegalCopyright\", STRINGIFY(${RC_COPYRIGHT})
+          VALUE \"LegalCopyright\", STRINGIFY(${ARG_COPYRIGHT})
         }
     }
     BLOCK \"VarFileInfo\"
@@ -29,6 +32,7 @@ VS_VERSION_INFO VERSIONINFO
     }
 }")
 
-set(_rc_file ${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}_res.rc)
-file(WRITE ${_rc_file} ${_rc_content})
-target_sources(${PROJECT_NAME} PRIVATE ${_rc_file})
+    set(_rc_file ${CMAKE_CURRENT_BINARY_DIR}/${target}_res.rc)
+    file(WRITE ${_rc_file} ${_rc_content})
+    target_sources(${target} PRIVATE ${_rc_file})
+endmacro()
