+++ include/opendspx/audioclip.h
@@ -1,0 +1,12 @@
+#ifndef OPENDSPX_MODEL_AUDIOCLIP_H
+#define OPENDSPX_MODEL_AUDIOCLIP_H
+
+namespace QDspx {
+
+    struct AudioClip {
+        
+    };
+
+}
+
+#endif //OPENDSPX_MODEL_AUDIOCLIP_H
