+++ include/opendspx/label.h
@@ -1,0 +1,15 @@
+#ifndef OPENDSPX_MODEL_LABEL_H
+#define OPENDSPX_MODEL_LABEL_H
+
+#include <QString>
+
+namespace QDspx {
+
+    struct Label {
+        int pos{0};
+        QString text;
+    };
+
+}
+
+#endif //OPENDSPX_MODEL_LABEL_H
