+++ include/opendspx/master.h
@@ -1,0 +1,14 @@
+#ifndef OPENDSPX_MODEL_MASTER_H
+#define OPENDSPX_MODEL_MASTER_H
+
+#include <opendspx/buscontrol.h>
+
+namespace QDspx{
+
+    struct Master {
+        BusControl control;
+    };
+
+}
+
+#endif //OPENDSPX_MODEL_MASTER_H
