+++ include/opendspx/singingclip.h
@@ -1,0 +1,12 @@
+#ifndef OPENDSPX_MODEL_SINGINGCLIP_H
+#define OPENDSPX_MODEL_SINGINGCLIP_H
+
+namespace QDspx {
+
+    struct SingingClip {
+        
+    };
+
+}
+
+#endif //OPENDSPX_MODEL_SINGINGCLIP_H
