+++ include/opendspx/tempo.h
@@ -1,0 +1,13 @@
+#ifndef OPENDSPX_MODEL_TEMPO_H
+#define OPENDSPX_MODEL_TEMPO_H
+
+namespace QDspx {
+
+    struct Tempo {
+        int pos{0};
+        double value{120.0};
+    };
+
+}
+
+#endif //OPENDSPX_MODEL_TEMPO_H
