+++ include/opendspx/timeline.h
@@ -1,0 +1,20 @@
+#ifndef OPENDSPX_MODEL_TIMELINE_H
+#define OPENDSPX_MODEL_TIMELINE_H
+
+#include <QList>
+
+#include <opendspx/label.h>
+#include <opendspx/tempo.h>
+#include <opendspx/timesignature.h>
+
+namespace QDspx {
+
+    struct Timeline {
+        QList<Label> labels;
+        QList<Tempo> tempos;
+        QList<TimeSignature> timeSignatures;
+    };
+
+}
+
+#endif //OPENDSPX_MODEL_TIMELINE_H
