+++ include/opendspx/timesignature.h
@@ -1,0 +1,14 @@
+#ifndef OPENDSPX_MODEL_TIMESIGNATURE_H
+#define OPENDSPX_MODEL_TIMESIGNATURE_H
+
+namespace QDspx {
+
+    struct TimeSignature {
+        int index{0};
+        int denominator{4};
+        int numerator{4};
+    };
+
+}
+
+#endif //OPENDSPX_MODEL_TIMESIGNATURE_H
