+++ include/opendspx/workspace.h
@@ -1,0 +1,13 @@
+#ifndef OPENDSPX_MODEL_WORKSPACE_H
+#define OPENDSPX_MODEL_WORKSPACE_H
+
+#include <QMap>
+#include <QJsonObject>
+
+namespace QDspx{
+
+    using Workspace = QMap<QString, QJsonObject>;
+
+}
+
+#endif //OPENDSPX_MODEL_WORKSPACE_H
