+++ include/opendspxconverter/opendspxconverterglobal.h
@@ -1,0 +1,18 @@
+#ifndef OPENDSPX_CONVERTER_OPENDSPXCONVERTERGLOBAL_H
+#define OPENDSPX_CONVERTER_OPENDSPXCONVERTERGLOBAL_H
+
+#include <QtCore/QtGlobal>
+
+#ifndef OPENDSPX_CONVERTER_EXPORT
+#  ifdef OPENDSPX_CONVERTER_STATIC
+#    define OPENDSPX_CONVERTER_EXPORT
+#  else
+#    ifdef OPENDSPX_CONVERTER_LIBRARY
+#      define OPENDSPX_CONVERTER_EXPORT Q_DECL_EXPORT
+#    else
+#      define OPENDSPX_CONVERTER_EXPORT Q_DECL_IMPORT
+#    endif
+#  endif
+#endif
+
+#endif //OPENDSPX_CONVERTER_OPENDSPXCONVERTERGLOBAL_H
