+++ include/opendspxserializer/opendspxserializerglobal.h
@@ -1,0 +1,18 @@
+#ifndef OPENDSPX_SERIALIZER_OPENDSPXSERIALIZERGLOBAL_H
+#define OPENDSPX_SERIALIZER_OPENDSPXSERIALIZERGLOBAL_H
+
+#include <QtCore/QtGlobal>
+
+#ifndef OPENDSPX_SERIALIZER_EXPORT
+#  ifdef OPENDSPX_SERIALIZER_STATIC
+#    define OPENDSPX_SERIALIZER_EXPORT
+#  else
+#    ifdef OPENDSPX_SERIALIZER_LIBRARY
+#      define OPENDSPX_SERIALIZER_EXPORT Q_DECL_EXPORT
+#    else
+#      define OPENDSPX_SERIALIZER_EXPORT Q_DECL_IMPORT
+#    endif
+#  endif
+#endif
+
+#endif //OPENDSPX_SERIALIZER_OPENDSPXSERIALIZERGLOBAL_H
