+++ src/converter/midi.cpp
@@ -1,0 +1,1 @@
+#include <opendspx/model.h>
