--- src/converters/opensvip.cpp
@@ -1,14 +1,0 @@
-#include "converters/opensvip.h"
-
-namespace QDspx {
-
-    Result OpenSvipConverter::load(const QString &path, Model *out, const QVariantMap &args) {
-        return {};
-    }
-
-    Result OpenSvipConverter::save(const QString &path, const Model &in,
-                                       const QVariantMap &args) {
-        return {};
-    }
-
-}
